#2480; adds amazonKeys (S3) state initialization.
